# Squash indirection layers on the key comparison path