Summary
========
`rocksdb` has multi-level abstraction over key comparison. Quick view on burnt cycles distribution with `perf` sampling from PMU:

```
-   26,91%     0,02%  db_bench     db_bench             [.] rocksdb::MemTableInserter::PutCF
     26,89% rocksdb::MemTableInserter::PutCF
      - rocksdb::MemTableInserter::PutCFImpl
         - 26,26% rocksdb::MemTable::Add
            - 24,96% rocksdb::InlineSkipList<rocksdb::MemTableRep::KeyComparator const&>::Insert<false>
               - 23,46% rocksdb::InlineSkipList<rocksdb::MemTableRep::KeyComparator const&>::RecomputeSpliceLevels
                  - rocksdb::InlineSkipList<rocksdb::MemTableRep::KeyComparator const&>::KeyIsAfterNode
                     - 9,10% rocksdb::MemTable::KeyComparator::operator()
                        - 4,60% rocksdb::InternalKeyComparator::Compare
                           - rocksdb::(anonymous namespace)::BytewiseComparatorImpl::Compare
                                __memcmp_sse4_1
                          1,80% rocksdb::GetVarint32PtrFallback
               + 1,10% rocksdb::InlineSkipList<rocksdb::MemTableRep::KeyComparator const&>::KeyIsAfterNode
              0,56% rocksdb::(anonymous namespace)::SkipListRep::Allocate
```

The indirection might be costly because of prohibiting compiler from inlining the code while imposing unnecessary memory dereferences. The changeset tries to deal with that through:
* just shuffling the `rocksdb::InternalKeyComparator::Compare` family from `.cc` to header,
* unifying `rocksdb::MemTableRep::KeyComparator` and `rocksdb::MemTable::KeyComparator` via common `rocksdb::KeyComparator`.

With the patches applied:
```
-   27,94%     0,06%  db_bench     db_bench             [.] rocksdb::MemTableInserter::PutCF
     27,88% rocksdb::MemTableInserter::PutCF
      - rocksdb::MemTableInserter::PutCFImpl
         - 27,07% rocksdb::MemTable::Add
            - 25,55% rocksdb::InlineSkipList<rocksdb::KeyComparator const&>::Insert<false>
               - 24,44% rocksdb::InlineSkipList<rocksdb::KeyComparator const&>::RecomputeSpliceLevels
                  - rocksdb::InlineSkipList<rocksdb::KeyComparator const&>::KeyIsAfterNode
                     - 2,94% rocksdb::(anonymous namespace)::BytewiseComparatorImpl::Compare
                          __memcmp_sse4_1
                       1,93% rocksdb::GetVarint32PtrFallback
                 0,74% rocksdb::InlineSkipList<rocksdb::KeyComparator const&>::KeyIsAfterNode
            + 0,70% rocksdb::(anonymous namespace)::SkipListRep::Allocate
```

Very rough performance comparison
==================

Before the rebase
----------------------
The branch started from pretty old `master` that is used in the Ceph project. It's still available as [wip-inlineable-KeyComparator-ceph](https://github.com/rzarzynski/rocksdb/tree/wip-inlineable-KeyComparator-ceph). Run under profiler, huge keys.

#### Reference point: 8dd0a7e11abad6bc32e59a5cba8769961e085312.

```
$ perf record -e cycles:pp --call-graph dwarf ./db_bench --writes 1000000 --benchmarks="fillrandom" --compression_type none -key_size 256

(...)

Initializing RocksDB Options from the specified file
Initializing RocksDB Options from command-line flags
RocksDB:    version 5.8
Date:       Thu Feb 15 22:09:31 2018
CPU:        8 * Intel(R) Core(TM) i7-6820HQ CPU @ 2.70GHz
CPUCache:   8192 KB
Keys:       256 bytes each
Values:     100 bytes each (50 bytes after compression)
Entries:    1000000
Prefix:    0 bytes
Keys per prefix:    0
RawSize:    339.5 MB (estimated)
FileSize:   291.8 MB (estimated)
Write rate: 0 bytes/second
Read rate: 0 ops/second
Compression: NoCompression
Memtablerep: skip_list
Perf Level: 1
------------------------------------------------
Initializing RocksDB Options from the specified file
Initializing RocksDB Options from command-line flags
DB path: [/tmp/rocksdbtest-1000/dbbench]
fillrandom   :       3.787 micros/op 264046 ops/sec;   89.6 MB/s

(...)

[ perf record: Captured and wrote 165.024 MB perf.data (20480 samples) ]
```

#### Changed: 5faa1dc9aa53d909bcfdee05be588803784f831f:
```
$ perf record -e cycles:pp --call-graph dwarf ./db_bench --writes 1000000 --benchmarks="fillrandom" --compression_type none -key_size 256

(...)

Initializing RocksDB Options from the specified file
Initializing RocksDB Options from command-line flags
RocksDB:    version 5.8
Date:       Fri Feb 16 00:12:59 2018
CPU:        8 * Intel(R) Core(TM) i7-6820HQ CPU @ 2.70GHz
CPUCache:   8192 KB
Keys:       256 bytes each
Values:     100 bytes each (50 bytes after compression)
Entries:    1000000
Prefix:    0 bytes
Keys per prefix:    0
RawSize:    339.5 MB (estimated)
FileSize:   291.8 MB (estimated)
Write rate: 0 bytes/second
Read rate: 0 ops/second
Compression: NoCompression
Memtablerep: skip_list
Perf Level: 1
------------------------------------------------
Initializing RocksDB Options from the specified file
Initializing RocksDB Options from command-line flags
DB path: [/tmp/rocksdbtest-1000/dbbench]
fillrandom   :       3.327 micros/op 300584 ops/sec;  102.1 MB/s

(...)

[ perf record: Captured and wrote 143.288 MB perf.data (17815 samples) ]
```

The rebased branch
--------------------------
Slightly longer runs without `perf record`.

#### Reference point: c88c57cde10a820b21d9fab0ca3f059da3852057.
```
$ ./db_bench --writes 20000000 --benchmarks="fillrandom" --compression_type none -key_size 256
Initializing RocksDB Options from the specified file
Initializing RocksDB Options from command-line flags
RocksDB:    version 5.10
Date:       Fri Feb 16 02:56:54 2018
CPU:        8 * Intel(R) Core(TM) i7-6820HQ CPU @ 2.70GHz
CPUCache:   8192 KB
Keys:       256 bytes each
Values:     100 bytes each (50 bytes after compression)
Entries:    1000000
Prefix:    0 bytes
Keys per prefix:    0
RawSize:    339.5 MB (estimated)
FileSize:   291.8 MB (estimated)
Write rate: 0 bytes/second
Read rate: 0 ops/second
Compression: NoCompression
Memtablerep: skip_list
Perf Level: 1
------------------------------------------------
Initializing RocksDB Options from the specified file
Initializing RocksDB Options from command-line flags
DB path: [/tmp/rocksdbtest-1000/dbbench]
fillrandom   :       3.779 micros/op 264615 ops/sec;   89.8 MB/s
```

#### Changed: 7a7f60a24c7007adc9a4af995cb722c1e16d7875
```
$ ./db_bench --writes 20000000 --benchmarks="fillrandom" --compression_type none -key_size 256
Initializing RocksDB Options from the specified file
Initializing RocksDB Options from command-line flags
RocksDB:    version 5.10
Date:       Fri Feb 16 02:41:39 2018
CPU:        8 * Intel(R) Core(TM) i7-6820HQ CPU @ 2.70GHz
CPUCache:   8192 KB
Keys:       256 bytes each
Values:     100 bytes each (50 bytes after compression)
Entries:    1000000
Prefix:    0 bytes
Keys per prefix:    0
RawSize:    339.5 MB (estimated)
FileSize:   291.8 MB (estimated)
Write rate: 0 bytes/second
Read rate: 0 ops/second
Compression: NoCompression
Memtablerep: skip_list
Perf Level: 1
------------------------------------------------
Initializing RocksDB Options from the specified file
Initializing RocksDB Options from command-line flags
DB path: [/tmp/rocksdbtest-1000/dbbench]
fillrandom   :       3.447 micros/op 290073 ops/sec;   98.5 MB/s
```