@ajkr 2019-01-03
I compared this against today's master branch to see if we're still missing out on a potential improvement. It looked like a pretty small difference in my contrived experiment.

setup command:
```
$ TEST_TMPDIR=/data/compaction_bench ./db_bench -benchmarks=fillrandom,readrandom -num=10000000 -write_buffer_size=1048576 -target_file_size_base=1048576 -max_bytes_for_level_base=4194304 -compression_type=none -max_background_jobs=12
-key_size=256 -value_size=64
```

benchmark command:
```
$ TEST_TMPDIR=/data/compaction_bench ./db_bench.pre -use_existing_db=true -benchmarks=readrandom -num=10000000 -threads=32 -compression_type=none -max_background_jobs=12 -key_size=256 -value_size=64
```

results before this PR:
```
readrandom   :       0.393 micros/op 2545649 ops/sec;  500.0 MB/s (10000000 of 10000000 found)
```

results after this PR:
```
readrandom   :       0.391 micros/op 2557407 ops/sec;  502.3 MB/s (10000000 of 10000000 found)
```